docs(asset-search): rewrite stories in new format (#544)
