--- src/components/AssetSearch/stories/basic.md
@@ -1,26 +1,0 @@
-## Basic 
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { AssetSearch } from '@cognite/gearbox';
-import { Asset } from '@cognite/sdk';
-
-function ExampleComponent(props) {
-  const onLiveSearchSelect = (item: Asset): void => {};
-  const strings = { searchPlaceholder: 'Try to type name of asset', }
-
-  return (
-    <AssetSearch
-      onLiveSearchSelect={onLiveSearchSelect}
-      strings={strings}
-    />
-  );
-
-}
-```
